Error en un casting, y error out of bounds exception
Quickfix

+++ src/Op.java
@@ -69,7 +69,7 @@ public class Op implements Calculadora {
 			}
 		}
 	
-		double respuesta = (double) datos.pop();
+		double respuesta = (int) datos.pop();
 		return respuesta;
